Better diagnostics in Test::is for strings differing only in whitespace
If the expected value differs from the got value only by whitespace at
the end of a line, it's very hard to find this difference. Using .perl
for the diag output on the other hand makes it trivial to find
whitespace differences.

With this patch is("foo \nbar baz", "foo\nbar baz"); gives:
expected: "foo\nbar baz"
     got: "foo \nbar baz"