Fix some issues with &hyper and optimize it
Fixes issues with hyper on empty lists and fixes some issues with hyper not
returning the right type. Prevented flattening behavior.

Additionally, the restructuring of &hyper cut up to 50% of the time it
took to run simple benches like:

    ((1, [2, 3]) «+» (4,5,[6,7],[8,9])) for ^1000

Also, this commit adds .Hash and .List methods to Any so that we can
coerce to them.