Delegate List() to List.new()
While we do have an Any.List coercer,
t/spec/integration/advent2013-day08.t has a case of

    class Vector is Array { … }

with the same semantics as Array, but we're missing a coercer for
Vector that &hyper can use. Since ListType.new() generally takes
a slurpy list, just like ListType() should, it makes sense for ListType() to
delegate to ListType.new(). And more importantly, it gets us passing
this test :)