use more [] and {} in .perl output for clarity
We now decont by suffix, but try to get the "what is this" information
out front, conveyed in this case by the brackets.  No more having to
read to the end of the list to see .hash.  For round-tripping purposes,
fixing this also entails making the .of types of [] and {} Mu rather than
Any, which is more consistent with declared arrays and hashes anyway.
We also have to parameterize Enum.perl to know if the intent is code
for an arglist, in which case we have to be more carefule about pair notation.