Handle exceptions from BEGIN blocks better.
We now include details of where the BEGIN block was in the code, the
fact that the error occurred while evaluating a BEGIN block, and a
stack trace inside of the BEGIN block. The mechanism will also be
useful for improving errors from the MOP (which we'll not show the
stack trace for) and other things like constants.