Make try do "use fatal", as per spec.
This makes various things fail more violently than they used to, so
we get a bunch of spectest failures as a result. We use a dynvar for
conveying if "use fatal" is in effect, which is consistent with S04's
requirement that "The fail function responds to the caller's use fatal
state. It either returns an unthrown exception, or throws..." Clearly
this has to be caller in the transitive to be of any use. Placing this
in a branch so we don't lose it.