Gracefully handle .gist on unthrown exceptions
If an exception is created and not thrown, and gist is called on it, the backtrace op is called without a VM-level exception object, and a long backtrace about not being able to make a backtrace is produced and appended to the gist.  This patch makes .gist simply print the Exception's message if it is called before the exception is thrown.