Re-implement 'use fatal' to be more correct.
Per TimToady++, a dynamic variable is the wrong approach. Rather, we
should do it by AST re-writing in the leixcal scope where 'use fatal'
is in force. This is a first attempt at implementing that. It's good
enough to pass all of the spectests that the previous 'use fatal'
implementation did. $*FATAL goes away as a result.