Support additional arguments to IO::Handle.open
In addition to the (fixed) :rw, there's now also :wr.
The former updates the file, the latter truncates it first.

The method now takes an optional mode descriptor as first argument.
Supported values are Perl 5's '<', '>', '>>', '+<', '+>', '+>>'
as well as the arguments to fopen() according to the C11 spec.

As-is, this commit breaks any unpatched backend.