Add support for additional mode flags to IO::Handle.open
In addition to :r, :w, :a and the (previously ignored) :rw,
there's now also :ra, :wx, :wr, :wrx, :ar.

With the exception of :ra (added for consistency), these correspond
to the modes of fopen() as specified by C11 - you get to the Perl6
name by replacing '+' with 'r'.

The $llmode parameter to the private !open method translates directly
to POSIX flags:

    r  O_RDONLY
    -  O_WRONLY
    +  O_RDWR
    a  O_APPEND
    c  O_CREAT
    x  O_EXCL
    t  O_TRUNC

'w' is a special compatibility alias that translates to '-ct' on its
own and to '-c' as part of a string. This ensures that 'w' and 'wa'
still map to their expected values.

Adding Perl5 into the mix, we end up with the following:

     p6      c11     moar    compat  p5
    -------------------------------------
     :r      r       r               <
     :rw     r+      +               +<
     :ra             +a
     :w      w       -ct     w       >
     :wx     wx      -cx     wx
     :wr     w+      +ct             +>
     :wrx    w+x     +cx
     :a      a       -ca     wa      >>
     :ar     a+      +ca             +>>

On the JVM, unsupported modes will blow up, but everything that used
to work should continue to work (with the caveat that :rw now blows
up instead of silently getting adjusted to :w).

On Parrot, unsupported modes will lead to undefined behaviour as
unknown characters get silently dropped from mode strings.