Simplify and remove race in role handling.
Since this code was written, it seems we (for other reasons) changed
the order things happen when setting up parametric roles and groups
of them. Therefore, simplify it. This also removes a race condition
if two threads tried to specialize the role at the same time, which
led to candidates maybe being added twice.