add 'unit' declarator
Unit scope declarations will in the future require a 'unit' declarator,
instead of relying merely on a blockless semicolon form.  (A semicolon
in place of a block is both too invisible and too late for such a feature.)
The absence of a 'unit' declarator on unit declarations is deprecated.

The identifier 'SemicolonForm' is now replaced with 'UnitScope'.