serialize CURLI database to "bytecode" instead of json
This has several reasons. The most important one is that this let use get
rid of the performance penalty which json gave us. When hundreds of dists
were installed the startup time increased to over 5s easily.
Now the startup time for -e1 is unchanged (lazy deserialization ftw), and
using (and therefore locating) modules is more effective. Also a benefit
of not having to deal with json is that we can stick Distribution and
CompUnit objects in our database, which reduces our CURL code a lot.