promote last, next and redo to full multi subs
at some point it had to be changed into this, but
it seems like it's safe to have it like this now.

the no-arguments version of those will now not only not
allocate a BOOTArray and a Parcel, but also get jitted
properly due to not using param_sp (slurpy positional)
op codes.