Don't run exit handlers on fake frames.
When we are seeing if a multi can pass the bind check, we create a
frame to hold lexicals used in the binding checking. This frame is
never actually entered and not fully set up, however. If the block in
question had exit handlers, we would end up looking at uninitialized
memory in trying to run them, which led to all kinds of oddness when
a frame with exit handlers (LEAVE, temp, etc.) had a signature with
where constraints. Fixes RT #124318.