Remove samewith() usage from core
Looking at how samewith() works, it felt better to not use in the core
from a performance pov.  Also, maybe samewith() should be implemented
as a macro, or perhaps as a statement in the grammar.  Or maybe not at
all.  In any case, it still feels wrong having to specify the name of
the method inside the method in general, so maybe we do need something
like samewith() (but better performing).