fix @foo = $bar = 5, 10;
Assignment is right associative, so the intent is that the righthand
assignment pay attention only to it's immediate left arg as its left term.
It would be incorrect to pay attention to treat the entire lefthand assignment
as the target of the righthand assignment.

The patch makes infix:sym<=> consume any $*LEFTSIGIL it finds, freeing up
the next term to set its own $*LEFTSIGIL.