Revive Proc::Status
This is a bandaid.  Some recent changes made Proc::Status disappear from
the world when still having code defining it.  Suspect some sort of MOP
messup with creating parent packages.  By creating a stub for the parent
package beforehand, we seem to have quelched the issue.  Something that
needs to be investigated further!