Disallow \$ \@ \% \& to create a Capture
For people coming from Perl 5, it is all too common to use \@a to
indicate a reference to @a.  In Perl 6, we don't do references as
visible as that.  But since using \ like that *does* do something
like it, (well actually a lot more), a migrating developer may get
the wrong idea about what it actually does, and why it is needed.
So I thought it would be a good idea to nip this in the bud by
completely disallowing it altogether.

Should you really want to create a Capture on a single scalar,
array, hash or sub, then you just need to add parentheses.  You
could argue that the create Capture operator is now \() instead
of just \.

This patch bluntly states "This looks like Perl 5 code" as a proof
of concept.  If consensus can be reached on this change, we can
create a better, more informative message.

At least 6 test-files in roast were using \ in the P5 or some other
incorrect sense, and passing tests for the wrong reason.  These
have been fixed.  The following test-files won't compile now, but
I think they contain a lot of bogus tests (from 2008) that may need
to be weeded out anyway:
  t/spec/S03-operators/eqv.t
  t/spec/S03-operators/value_equivalence.t

Please revert if so needed.