Start preparing for multi-dim *-POS.
We'll pass multiple indices to the [AT|ASSIGN|EXISTS...]-POS methods.
This will allow things that know how many dimensions they have to just
bind them directly to parameters, which will be efficient (and allow
the dynamic optimizer to inline, etc.). We'll be able to do some case
analysis on the indexer itself at optimization time to eliminte the
LoL in the commonest cases also.

This patch simply adds fallbacks for the multi-indice case that dig
into the data structure a level at a time.