Test for concise error on sinking last statement.
If the last statement in a file used to need sinking, and doing so
threw an exception, we'd spit out an epic low-level backtrace rather
than a nice high-level one. Fix that by making non-EVAL and non-REPL
stuff we compile put the last statement in sink context in the code we
generate rather than having MAIN do the sinking. Also means a trailing
0; in a file now warns.