Straighten out what goes in Any and List.
A couple of things that don't need a reified List to reasonably work
(unique, squish, reduce) have had their implementation moved into the
Any-list-methods. Those left in List are things that really do need
all the values to work (permutations, combinations, etc.). The sort
method is an interesting case; it needs all the values, but it needs
to get them into something fresh that it can mutate. Further, we may
want a parallelizing version of it some day for HyperIterable. So it
goes in Any.