Lift Routine's $!why behavior up to Block
There's no reason that Blocks shouldn't be able to be doc'd.  The
inspiration for this was creating a little Bailador application:

    #| Get the given user
    get '/user/:user', {
        ...
    };

Bailador, other web frameworks, or any module that provides a blocky
sort of DSL could really take advantage of declarative docs on blocks;
for example, to generate documentation HTML for the routes in an
application.