Begin fixing up reductions.
Make the code-gen respect the "single item" rule we're using in many
places, and re-implement left-assoc reduce to work this way. Note that
it uses the iterator API directly and should be a good bit faster than
what came before, which shuffled things in and out of a temporary
intermediate list.