Remove spurious laziness checks.  Please.
It is neither specced nor tested for that an unshift to a lazy
construct should be illegal.  There are commented out tests
in roast, but they are obviously just copies from when someone
cut and pasted push-related code to make the unshift test file,
and it is quite likely that the actual tests in current code is
likewise just leftover from not noticing that this test is not
needed for unshift.

(It is specced that *infinite* things fail when they are the
things being unshifted onto the target, but this is not
implementable unless we have some sort of depth limiter pragma.)