Add naive implementation of Iterator.count-only
The idea being that if we have a situation like "foo".IO.lines.elems, we
can call "count-only" on the iterator, which would not bother pushing anything
to the target, but do a much faster low level count.  *AND* would allow any
consumer class to optimize that case even further (like now with the :count
adverb on IO::Handle.lines).