Add tentative `react` block.
The supply block is great for processing stages, but you'll often have
some kind of "top level" in an asynchronous application - or those new
to the Perl 6 asynchronous features will want an easy way to play with
things. Creating a Supply to immediately wait on it would thus be a
really common pattern. `react { ... }` does that, warns on bogus emit
usages, and throws any unhandled async exception in the react.