Make push and unshift iterate on single arguments in containers.
Per http://irclog.perlgeek.de/perl6/2015-08-24#i_11106557
itemization only protects against flattening, not single-arg-rule.

When push and shift were implemented they assumed the opposite.

This brings their behavior in line and fixes as much of the
fallout from that in rakudo.  Even with those fixes we lose
progress on the test suite -- 7 files and some more tests in
already failing files need adjustment because they relied on
the previous behavior.

Some of the internal fixes may be superstitious, but can be
cleaned up later (you must be able to know you will not get any
Iterables to determine what is safe, and that is time consuming.)