Make "done" a full supply citizen
Tomorrow, Jonathan is going to give yet another installment of his latest
async presentation.  We want people to use the new goodies as soon as
possible and not have to change their async code again.  So I thought it
would be a good idea to rename supply-emit and supply-done to just "emit"
and "done".  This then interferes with the "done" from Test.pm (which I
recently deprecated).  It seemed like a good time, during this upheaval,
to at least make this work in the GLR branch.  So "done" is eradicated from
Test.pm (please use "done-testing" instead if you must: you should probably
use "plan" anyway).