Add allomorphic types
This introduces IntStr, RatStr, NumStr, and ComplexStr. I wasn't sure of
how to implement them exactly, since the spec is pretty sparse on it,
but I went about it by thinking that most things should let MMD decide
how to view the types, so only a minimal number of methods have been
implemented here.

The choice of .gist is debatable, I just wanted something that showed
off it was an allomorphic type, but not quite as "technical" as .perl is
supposed to be. Alternatives to .gist welcome, however :) .

I also just noticed that the constructor as shown in S02 takes one
argument; these take two, on the assumption that val() will do all the
work. However, if those constructors were changed to their spec'd "just
a string" version, then perhaps the val() I wrote (coming up next
commit) could be split up and made a bit more manageable.

Finally, this necessitated a new src/core file, which felt like the most
appropriate move, and the least disruptive (compared to the alternative
of moving src/core file order around). It's just before Main.pm, since
AFAIK sub MAIN is the only thing in src/core that's specifically meant
to use val() (and it currently has a workaround for it in there).