Add the val() function.
Takes a Str, returns an allomorphic type if possible, or failing that
the Str. This massive function's structure was only vaguely inspired by
Str.Numeric, basically just reusing the use of inner subs in the
function, for those curious :P . Even then it was written separately
from other implementations of string-to-numeric stuff.

Isn't quite perfect, for example don't try ":16<F*:2<1*2**10>**2>" or
":16<F*2**:2<10>>". Those sorts of things are pretty much asking for a
regex-based implementation of val(), but that was avoided here based on
the idea that using a mini-grammar would be slower than desirable.
However, use of a grammar (or $~MAIN's various grammar rules, if
possible) should be considered anyway, and would likely be a lot clearer
than this performance-considering version.

The large pod comment prefacing the function is an attempt to help
clarify how all the inner subs relate to each other, to aid
understanding for both myself and others. I'd be fine with moving it to
a separate file in rakudo/docs, for example, but I don't think it should
be removed completely. It and the declarator blocks on the inner subs
have been commented out since the Pod portion of the parser dies on
them while processing CORE.