Make multiple Q adverbs work.
This involved quite a bit of changes to how the post-processing adverbs
are handled. They can now be mixed appropriately (though :to is
currently limited to being the last post-processing adverb). A slurpy
multi was added for val(), as that appeared to be the easiest way to
handle q:w:v// . Doesn't stop you from duplicating adverbs, e.g.
q:v:v:v//, but that's hopefully not an issue.