Make sure that @arr.List creates a completely disconnected List
@arr.clone creates a shallow clone that contains the same $!reified
buffer, so changes to the original would change the clone as well.

Just using List.from-iterator(self.iterator) would ensure that we get a
fresh $!reified buffer, but the items still share the containers. So
when we assign to the array items, we still change the list's items as
well.

We're now deconting the Array's values when pushing them into the new
List's $!reified. This makes $list.Array.List a bit more symmetrical,
since .Array itemizes all values. However, if the values were items to
begin with, we now decont them anyway. So it's not completely
symmetrical, but should hopefully give fewer surprises overall.