Replace Str.Numeric with val() usage
This is in part to reduce code duplication, and in part because it feels
more natural (at least to me) to have string numification use val();
that's what the function is meant to do, after all :) . Exposes a couple
of blind spots in val(), particularly Inf and NaN. Also doesn't convert
"" to 0, but I happen to think that's wrong in the first place.

Also, it would be nice for val() to offer a better error reporting for
Str.Numeric's use. That would probably require a :val-or-fail adverb to
val() or something.