Make val() offer decent error reporting for Str.Numeric
Made possible via an adverb to just the single-arg Str multi (since
that's what Str.Numeric uses). Various changes involved include
rearranging choices handed to &try-possibles to resemble something
vaguely LTM-like, and a &has-to-be-this that works along with
&try-possibles in a fashion similar to :: or similar in regexes.

And yes, if you haven't caught on yet, val() is essentially doing what a
grammar would do, but less readable and/or concise. I don't know if a
grammar-based val() would be faster or slower than what's currently
here, though.