Experimental support for assigning to slices of Seqs (of containers)
Allow an assignment operation such as this to "work":

eager @a.map(-> $l is rw { $l })[0,1] = 3,4;

Note the eager, because Seq assignment is lazy.  To get the assigment to
actually work, it must be there, or the leftmost thing in a chain of
assignments must be Positional.  Laziness is emphasized to the extent
that self-reference on the right side or inside the slicer may be unintuitive.

The semantics of .[] on Seqs is not well defined yet so expect flux.