Add a check in val() to quicken some calls.
This shortcuts the call for strings with absolutely no digits in them,
by saying they can't possibly be numbers. Only 'Inf' and 'NaN' need to
be special-cased (as well as 'i', though that never worked in
numification, so I'm not sure of that). This shortens a particularly
long sorting operation from 90s to around 25s for me.