Reinstate the string-to-number code by japhb++
It's a bit faster than my version, and it was the original
implementation, so in it goes, in preference of me wasting time trying
to optimize my version further (since such optimizations at this point
would've been of the trickier, more subtle variety).

The biggest functional change to it, aside from making it work in val(),
was disallowing 'Infi' and 'NaNi' as valid numbers. Other test failures
as a result of this code concern things like if Rat fractions should
allow negative denominators, so it's been left untouched for now.

Unfortunately, it seems that the particular implementation of val() is
not the cause for the perceived slowdown anymore, so this replacement
doesn't offer much improvement. The cause is likely the existence of a
bunch of val() calls on quotewords in CORE existing now. Handling val()
calls at compile time where possible might offer the best improvement at
this point.