Implement a compile-time version of val() in World
This is called to help streamline val() processing on strings with
compile_time_values. This version, unlike the one in src/core, calls on
the Perl 6 grammar to do the work (something val() might like to do
someday). It also has three possible return values: an allomorphic
constant, if possible, the original string QAST if we know it can't be a
value, or a call to val() where we're unsure of the string.

This takes my sorting benchmark from ≈27.7s to ≈16.2s average, compared
to nom's 12.4s average. This seems to me to be sufficiently close to
pre-val() speeds that this could be merged into nom whenever
appropriate.

This test passes all spectests the previous commit did (that is, those
tests relying on a stricter val() definition still fail, pending
decisions about allomorphic types in this area).