s/rw/raw/ on slurpies
It doesn't make sense to call 'is rw' on slurpies, since 'rw' is supposed
to guarantee a valid lvalue.  These are now just 'is raw' (what used to
be 'is parcel'), which is more in line with what is wanted.  In other words,
whether the arg is an lvalue or not depends entirely on whether you passed one.