Various improvements to infix:<X>.
Remove various superstitious bits of code that worked around the way
things were pre-GLR. Then make sure we .list the things we are going
to work over. It turns out that if a Range were one of the arguments
to infix:<X> then indexing it would be rather more expensive than if
we had done .list on it in the first place.