Change back allomorphic .perl & .gist
After thinking about it for a while, I decided to change back the .gist
and .perl methods for the allomorphic types to how they were originally.
We really need .gist to show that you're dealing with an allomorphic
type, and so it prints the use of val(), since that's the most succinct
way I know of to do it.

The .perl method shows how to construct the object via .new, since I
figure .perl is more about the object itself than whatever uses it. And
considering the allomorphic types are meant to be user-facing (esp. if
you use something like MAIN), I don't think it's so terrible to show it
for .perl.

For neither method is using quotewords syntax (<> and «») appropriate;
not all uses of val() are through those quoting syntaxes, and those uses
of val() with leading/trailing whitespace would lead to an incorrect
.perl previously. If we wanted to use quoting syntax for either or both
of these methods, it'd have to be a representation using q:v[] or
similar.