Make Whetever.new == Inf True
It occurred to me that we have many places in the core setting where we are
checking for whether a value is either Whatever or Inf.  It also occurred
to me that adding a Numeric method for Whatever would make such a check much
easier, because it would just allow you to say:

  sub a($value) { say "foo" if $value == Inf }

and pass either a(*) or a(Inf) would give the desired effect.  Before
applying this new capability, I thought I'd get it out in the world first.