handle META6.json files in CURLF
That means we finally can have full unicode package names put into the META6.json
in the provides section, but let that entry point to a file containing e.g. ASCII
characters only. If one does -I/directory where the META6.json is in there, the
mapping of package names to file paths will work out.
This does not affect the current way of including -I/directory/lib directly, as
this will not know about a META6.json.