GLRify IO::Handle.words properly
Makes things between 30% ("words".IO.words[{ $_-10 .. $_-1 }]) and 5x
faster ("words".IO.words.elems).

The limit positional variable has been removed: because of a proto issue,
this has never worked for IO.words and was not spectested.  It appears that
post-GLR this optimization is basically not needed, since it only marginally
makes things faster at the expense of serious code bloat.