List.ACCEPTS now uses ~~ and treats ** as glob
List.ACCEPTS was really wimpy and not pulling its weight.  It now
recursively matches a list with smartmatch.  Since * is a valid smartmatch
matching a single item, we now use ** as the globber that can match
multiple items, which is more consistent with the rest of the language.
And since the submatches are now approximate, we can get temporary false
positives, so in that case it now backtracks more like regex patterns do.