Add 'bool-only' method for Iterators
The idea being that if we want the boolean value of a Seq, we can call this
method (which would only do a pull-one once), and thus save a *lot* of work.

Making Seq.Bool call this method, only causes 1 test-file with errors:
t/spec/S03-smartmatch/signature-signature.t . Not sure whether that are
just lame tests or something more nefarious.  Waiting for a general ++ from
jnthn for the concept before digging deeper into this.