Add randomize option
The idea being that if test files are executed in random order, the ones that
are twiddling their thumbs (like the S17-supply tests) will be more evenly
distributed over the course of the test, and thus cause a lower wallclock
time.  So far, that didn't turn out the case just yet. :-(  Adding the option
for further testing anyway.