Reinstate original grep handler for junctions
The new Seq based code for some reason does not handle grepping using with
a junction (at least according to test 28 in S32-list/grep.t).  Adding this
candidate using the original code, makes the test pass.  Not 100% if that
test is legit, though.  At least this reinstates the previous behaviour.