One possible solution to RT#120895
Deindents delete from the front.  Any spaces that are not
contributing to the length are deleted along with the tab
they are concealed by.

Spaces from tab explosions still get added to the end of
the remaining whitespace.  You can no longer expect
e.g. .indent(4).indent(-4) to roundtrip, but this will
consistently cause tabs to be moved leftward and deleted
when mixing spaces.

The example in the RT will work if $?TABSTOP is set to 4 first.
(We can't preserve visual consistency without requiring this.)