Remove leftover parrot cruft from t/harness
This includes the icu option (since, as I recall, that was a Parrot
thing, and there are no # icu lines in spectest.data anymore anyway), as
well as the parrot-revision option and attempts to use perl6-p or fudge
on rakudo.parrot.

The ternaries that were modified simply default to moar now. I figured
that would be the safest option, though I could also see some sort of
exception being thrown in those spots instead, or something along those
lines.