Implement "slow" fudger
A test marked as "slow" will not be done in alphabetical order.  Instead,
all tests marked "slow" injected as far apart from each other in the total
run of testing.  When given enough TEST_JOBS (e.g. 10 on a quad-core CPU)
this results in almost 100% CPU usage for the whole run.  Which should
result in smaller wallclock times.