Allow infix:<.> if next thing is alphabetic
As much as I mistrust "fluent" programming, we already have an infix:<.=>
operator, so it's not a big stretch to have infix:<.>.  We limit it
to actual method names so that we can detect (most) accidental use
of P5's use of . for concatenation.  Note that this is a term/infix
distinction now, so .meth where a term is expected still defaults to $_.
Community standards for method cascades will likely ask for whitespace
around infix:<.> to make a visual disinction from the term form, but we
don't enforce that, if for no other reason than to refrain from giving
the fluent programmers something to yammer on about.  Basically, I've
decided that the slight possibility for greater confusion is outweighed
by consistency with .= and utility of have a . that can also serve as
a precedence modifier.  (Since both .= and . parse leftward with
item assignment precedence, but rightward as a term.)