. and .= should parse tighter
To be intuitive, . and .= need to parse at about the level of unaries,
such that "-3**2 . abs" treats the left as a single pill, but nothing
looser than that, lest it eat up the multiplication of "-3 * 2 . abs".
(Changing .= away from item assignment is okay because initializer .=
was effectively tighter precedence anyway, and nobody ever noticed.
And .= was always syntactically privileged anyway.)