Refactor outdent code
When writing spectests for 12db6c0 I noticed a corner case that was
not giving visually consistent results.  Also now that we delete
from the front the code was not arranged very logically.  Refactor
to take advantage of rx features and take substrings of the original
input rather than splitting/rejoining character-by-character.

The corner case was that a tab appearing in the uncut portion of
the whitespace would not move during an uneven dedent.  Now, enough
space is removed from the front to assure the remaining tabs notch
in one tabstop, then extra padding is added to the end to compensate
for their movement.