distinguish push/unshift from append/prepend
For one-arg semantics, we now use prepend and append.  The push and unshift
methods/subs are now designed to primarily support pushing or unshifting
a single item (without you having to explicitly itemize, which would
defeat slippage for things like the sequence operator).  So now you can
say @dest.push(@src.pop) without worrying about accidental flattening of
a single element that happens to be iterable.  Plus it will save wear and
tear on your $ key when you want to push a single array or list.