Implement bracket-delimited categorical sub declarations.
This fixes RT #119919. RT #117737 is related and still behaves the same, which
I would argue as correct in the first place - a bare regexp that matches a
single character is not a useful enough compile-time value to declare an
operator with. I've added a typed exception to be thrown in that case though.