handle Type:D and Type:U on variables and attributes
`my Int:D $foo` compiles to something like `my Int $foo where *.defined = 0`
now.  Some occurrences of Int:D in the setting had to be removed because at
the time it is compiled we can't instanciate Pair.  And we need Pair to add
the constraint to the variable or attribute.
However, these occurrences had no affect yet anyway.