Correct semantics of dieing LEAVEs.
Just in time for autumn! Now as per design docs, an exception thrown
by a LEAVE does not prevent evaluation of further LEAVE/UNDO/KEEP/POST
blocks. If we get one exception we just throw it; if we get multiple
we collect them all and throw an X::PhaserExceptions containg them.