Fix RT #125689.
For some reason the lookup with $*W.find_symbol in apply_trait doesn't look for
X::Inheritance::UnknownParent but Inheritance::UnknownParent in 'class X is
nosuchtrait { }', presumably because it happens 'inside' package X. Changing
the lookup to only look inside the setting loops infinitely when compiling the
setting.  The empty (but commented) CATCH prevents this.