generalize N-ary reduce/produce to [op] & [\op]
The N-ary code was only implemented in method reduce (and by extension,
sub reduce).  This is now moved over to the REDUCE generators in metaop.pm
so those also do N-ary correctly.

Additionally, the code should now work for functions that return lists and
not just for scalars.  (Not tested, but same transformation we used in the
sequence operator to use push (or unshift) that is overridable with a slip,
so a list is treated as a unit by default, but can be overridden.)