Make Mixy.roll return 1 value as specced
Please note that the code in Mix/MixHash is identical.  Originally, this
lived in the Mixy role.  This compiled allright, but gave runtime errors
("Cannot invoke this object") on the return statement.  Moving the class
to Rakudo::Internals caused the settings to fail to compile with "Cannot
find method 'compose'".  So, for now, the code is duplicated (except for
the invocant signature) in the two classes.