QX deprecated, runout and shellout
There are many reasons for :x adverb to go, most of them are explained
in the pull request. In short: out of other alternatives (namely "run")
qqx// and friends seem to be the worst ones (due to potential misuse and
associated security problems). In such situation it is appropriate to
get Huffman coding right, this commit does it.