await $supply works, can mix with Promise/Channel
Because it will now call $supply.await.  This may need some further
discussions.  To *me*, this feels like the natural thing, since
Supplies have their handling normally somewhere else anyway.  If
this is consensed to be a bad idea, then maybe it should not call
await, but the Promise method instead.