Supply.throttle control message is now key:value
Rather than using a Pair (which is awkward to use in an .emit anyway,
because you need extra parens to not make it a named parameter), we
now use a string of the form "key:value".  This should *also* make it
much easier to use Supply.throttle as a means of controlling supply
processing across multiple machines.